Workaround CentOS dependency conflicts
`/sbin/weak-modules` is present in the CentOS Docker image, so when we
try to un-install a Kernel package (we do so when cleaning things up,
the Kernel package itself is a very indirect dependency of git...), it
runs, and then complains that Dracut should be installed, but Dracut
itself depends on the Kernel, so it must be removed first, and things
fail.

This worksaround the issue by tricking the Kernel's uninstall scriptlet
into not calling `/sbin/weak-modules` (by moving it somewhere else
altogether - it'll ignore it if it's missing) so that it doesn't fail on
not finding Dracut.